Android Project
This Android Project is about a Math Tutor which can be used for Children. It includes Additions, Subtractions, Multiplications and Divisions and a combination of all. This application also has two Grading Calculators for two classes and a Mad Libs Application along with a Quiz application. This application also includes a Rock, Paper, Scissors game.